couldn't do it myself, modified someone elses implementation
credit to: https://albertauyeung.github.io/2020/06/15/python-trie.html

However... I think a trie is not enough... I need all unique
sub-sequences... I think that's a bit more of a combinatorial
question, all combinations of sequential pitches in a song. This
implies combinations of length 2 upto length N, where N is the number
of notes in a song... Maybe a better representation would be a graph.